chore(github): delete dead in-tree poller/board-sync + orphaned wiring (GH-4191)
