--- internal/adapters/github/poller_integration_test.go
@@ -1,580 +1,0 @@
-//go:build integration
-
-package github
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"strconv"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-)
-
-// TestPoller_Integration_IssueDiscovery verifies real Poller discovers issues correctly
-func TestPoller_Integration_IssueDiscovery(t *testing.T) {
-	// Track API calls
-	var mu sync.Mutex
-	apiCalls := make(map[string]int)
-
-	// Track which issues have been "processed" (would get pilot-done label)
-	processedInServer := make(map[int]bool)
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		mu.Lock()
-		apiCalls[r.URL.Path]++
-		mu.Unlock()
-
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			// Return issues with pilot label, but mark processed ones as done
-			mu.Lock()
-			issues := []*Issue{}
-			if !processedInServer[1] {
-				issues = append(issues, &Issue{
-					Number:    1,
-					Title:     "Test Issue 1",
-					Body:      "Test body 1",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now().Add(-2 * time.Hour),
-				})
-			}
-			if !processedInServer[2] {
-				issues = append(issues, &Issue{
-					Number:    2,
-					Title:     "Test Issue 2",
-					Body:      "Test body 2",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now().Add(-1 * time.Hour),
-				})
-			}
-			mu.Unlock()
-			json.NewEncoder(w).Encode(issues)
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	// Track processed issues
-	var processedIssues []int
-	var processMu sync.Mutex
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithOnIssue(func(ctx context.Context, issue *Issue) error {
-			processMu.Lock()
-			processedIssues = append(processedIssues, issue.Number)
-			processMu.Unlock()
-
-			// Mark as processed in mock server
-			mu.Lock()
-			processedInServer[issue.Number] = true
-			mu.Unlock()
-			return nil
-		}),
-		WithMaxConcurrent(1),
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	// Start poller in background
-	go poller.Start(ctx)
-
-	// Wait for issues to be processed
-	time.Sleep(500 * time.Millisecond)
-	cancel()
-
-	// Verify issues were discovered
-	processMu.Lock()
-	count := len(processedIssues)
-	processMu.Unlock()
-
-	if count != 2 {
-		t.Errorf("Expected 2 issues processed, got %d", count)
-	}
-
-	// Verify API was called
-	mu.Lock()
-	issuesCalls := apiCalls["/repos/test/repo/issues"]
-	mu.Unlock()
-
-	if issuesCalls < 1 {
-		t.Error("Expected at least 1 issues API call")
-	}
-}
-
-// TestPoller_Integration_LabelFiltering verifies label-based filtering
-func TestPoller_Integration_LabelFiltering(t *testing.T) {
-	var mu sync.Mutex
-	issue10Processed := false
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			mu.Lock()
-			issues := []*Issue{
-				// Only return issue 10 if not yet processed
-				{
-					Number: 11,
-					Title:  "Has pilot-in-progress label (should skip)",
-					State:  "open",
-					Labels: []Label{
-						{Name: "pilot"},
-						{Name: "pilot-in-progress"},
-					},
-					CreatedAt: time.Now(),
-				},
-				{
-					Number: 12,
-					Title:  "Has pilot-done label (should skip)",
-					State:  "open",
-					Labels: []Label{
-						{Name: "pilot"},
-						{Name: "pilot-done"},
-					},
-					CreatedAt: time.Now(),
-				},
-				{
-					Number: 13,
-					Title:  "Has pilot-failed label (should skip)",
-					State:  "open",
-					Labels: []Label{
-						{Name: "pilot"},
-						{Name: "pilot-failed"},
-					},
-					CreatedAt: time.Now(),
-				},
-			}
-			if !issue10Processed {
-				issues = append([]*Issue{{
-					Number:    10,
-					Title:     "Has pilot label",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now(),
-				}}, issues...)
-			}
-			mu.Unlock()
-			json.NewEncoder(w).Encode(issues)
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	var processedIssues []int
-	var processMu sync.Mutex
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithOnIssue(func(ctx context.Context, issue *Issue) error {
-			processMu.Lock()
-			processedIssues = append(processedIssues, issue.Number)
-			processMu.Unlock()
-
-			mu.Lock()
-			if issue.Number == 10 {
-				issue10Processed = true
-			}
-			mu.Unlock()
-			return nil
-		}),
-		WithMaxConcurrent(1),
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-	time.Sleep(500 * time.Millisecond)
-	cancel()
-
-	processMu.Lock()
-	defer processMu.Unlock()
-
-	// Only issue #10 should be processed (no status labels)
-	if len(processedIssues) != 1 {
-		t.Errorf("Expected 1 issue processed, got %d", len(processedIssues))
-	}
-
-	if len(processedIssues) > 0 && processedIssues[0] != 10 {
-		t.Errorf("Expected issue #10 to be processed, got #%d", processedIssues[0])
-	}
-}
-
-// TestPoller_Integration_SequentialMode verifies sequential execution
-func TestPoller_Integration_SequentialMode(t *testing.T) {
-	var issueOrder []int
-	var orderMu sync.Mutex
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			json.NewEncoder(w).Encode([]*Issue{
-				{
-					Number:    100,
-					Title:     "Oldest issue",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now().Add(-3 * time.Hour),
-				},
-				{
-					Number:    101,
-					Title:     "Middle issue",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now().Add(-2 * time.Hour),
-				},
-				{
-					Number:    102,
-					Title:     "Newest issue",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now().Add(-1 * time.Hour),
-				},
-			})
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithExecutionMode(ExecutionModeSequential),
-		WithOnIssueWithResult(func(ctx context.Context, issue *Issue) (*IssueResult, error) {
-			orderMu.Lock()
-			issueOrder = append(issueOrder, issue.Number)
-			orderMu.Unlock()
-			// Return success with no PR (direct commit simulation)
-			return &IssueResult{
-				Success:  true,
-				PRNumber: 0,
-				HeadSHA:  "abc123",
-			}, nil
-		}),
-		WithSequentialConfig(false, 100*time.Millisecond, 1*time.Minute),
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-	time.Sleep(1 * time.Second)
-	cancel()
-
-	orderMu.Lock()
-	defer orderMu.Unlock()
-
-	// In sequential mode, oldest issue should be processed first
-	if len(issueOrder) < 1 {
-		t.Error("Expected at least 1 issue to be processed")
-	}
-
-	if len(issueOrder) > 0 && issueOrder[0] != 100 {
-		t.Errorf("Expected oldest issue #100 first, got #%d", issueOrder[0])
-	}
-}
-
-// TestPoller_Integration_PRCallback verifies OnPRCreated callback
-func TestPoller_Integration_PRCallback(t *testing.T) {
-	var callbackCalled int32
-	var callbackPRNum int
-	var callbackIssueNum int
-
-	var mu sync.Mutex
-	issue200Processed := false
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			mu.Lock()
-			issues := []*Issue{}
-			if !issue200Processed {
-				issues = append(issues, &Issue{
-					Number:    200,
-					Title:     "Issue with PR",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now(),
-				})
-			}
-			mu.Unlock()
-			json.NewEncoder(w).Encode(issues)
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithExecutionMode(ExecutionModeSequential),
-		WithOnIssueWithResult(func(ctx context.Context, issue *Issue) (*IssueResult, error) {
-			mu.Lock()
-			issue200Processed = true
-			mu.Unlock()
-			return &IssueResult{
-				Success:    true,
-				PRNumber:   500,
-				PRURL:      "https://github.com/test/repo/pull/500",
-				HeadSHA:    "sha500",
-				BranchName: "pilot/GH-200",
-			}, nil
-		}),
-		WithOnPRCreated(func(prNumber int, prURL string, issueNumber int, headSHA string, branchName string, issueNodeID string) {
-			atomic.AddInt32(&callbackCalled, 1)
-			callbackPRNum = prNumber
-			callbackIssueNum = issueNumber
-		}),
-		WithSequentialConfig(false, 100*time.Millisecond, 1*time.Minute),
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-	time.Sleep(500 * time.Millisecond)
-	cancel()
-
-	// Verify callback was called at least once
-	if atomic.LoadInt32(&callbackCalled) < 1 {
-		t.Errorf("Expected OnPRCreated callback to be called at least once, got %d", callbackCalled)
-	}
-
-	if callbackPRNum != 500 {
-		t.Errorf("Expected PR number 500, got %d", callbackPRNum)
-	}
-
-	if callbackIssueNum != 200 {
-		t.Errorf("Expected issue number 200, got %d", callbackIssueNum)
-	}
-}
-
-// TestPoller_Integration_ParallelExecution verifies parallel mode with semaphore
-func TestPoller_Integration_ParallelExecution(t *testing.T) {
-	var concurrentCount int32
-	var maxConcurrent int32
-
-	var mu sync.Mutex
-	processedIssues := make(map[int]bool)
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			// Return multiple issues to trigger parallel processing
-			mu.Lock()
-			issues := []*Issue{}
-			for i := 0; i < 5; i++ {
-				if !processedIssues[300+i] {
-					issues = append(issues, &Issue{
-						Number:    300 + i,
-						Title:     "Parallel issue",
-						State:     "open",
-						Labels:    []Label{{Name: "pilot"}},
-						CreatedAt: time.Now().Add(-time.Duration(i) * time.Hour),
-					})
-				}
-			}
-			mu.Unlock()
-			json.NewEncoder(w).Encode(issues)
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithExecutionMode(ExecutionModeParallel),
-		WithOnIssue(func(ctx context.Context, issue *Issue) error {
-			// Track concurrent executions
-			current := atomic.AddInt32(&concurrentCount, 1)
-			for {
-				max := atomic.LoadInt32(&maxConcurrent)
-				if current > max {
-					if atomic.CompareAndSwapInt32(&maxConcurrent, max, current) {
-						break
-					}
-				} else {
-					break
-				}
-			}
-
-			// Simulate work
-			time.Sleep(100 * time.Millisecond)
-
-			// Mark as processed
-			mu.Lock()
-			processedIssues[issue.Number] = true
-			mu.Unlock()
-
-			atomic.AddInt32(&concurrentCount, -1)
-			return nil
-		}),
-		WithMaxConcurrent(2), // Allow 2 concurrent executions
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-	time.Sleep(2 * time.Second)
-	cancel()
-
-	// Wait for active tasks to complete
-	poller.WaitForActive()
-
-	// Verify concurrency was limited
-	max := atomic.LoadInt32(&maxConcurrent)
-	if max > 2 {
-		t.Errorf("Expected max concurrent <= 2, got %d", max)
-	}
-}
-
-// TestPoller_Integration_ProcessedStore verifies persistent store integration
-func TestPoller_Integration_ProcessedStore(t *testing.T) {
-	// Mock store that tracks processed issues
-	store := &mockProcessedStore{
-		processed: make(map[int]bool),
-	}
-
-	var mu sync.Mutex
-	issue400Processed := false
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/repos/test/repo/issues":
-			mu.Lock()
-			issues := []*Issue{}
-			if !issue400Processed {
-				issues = append(issues, &Issue{
-					Number:    400,
-					Title:     "Stored issue",
-					State:     "open",
-					Labels:    []Label{{Name: "pilot"}},
-					CreatedAt: time.Now(),
-				})
-			}
-			mu.Unlock()
-			json.NewEncoder(w).Encode(issues)
-		default:
-			w.WriteHeader(http.StatusOK)
-		}
-	}))
-	defer server.Close()
-
-	client := NewClientWithBaseURL("test-token", server.URL)
-
-	var processCount int32
-
-	poller, err := NewPoller(client, "test/repo", "pilot", 100*time.Millisecond,
-		WithOnIssue(func(ctx context.Context, issue *Issue) error {
-			atomic.AddInt32(&processCount, 1)
-			mu.Lock()
-			issue400Processed = true
-			mu.Unlock()
-			return nil
-		}),
-		WithProcessedStore(store),
-		WithMaxConcurrent(1),
-	)
-	if err != nil {
-		t.Fatalf("NewPoller failed: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-	time.Sleep(500 * time.Millisecond)
-	cancel()
-
-	// Issue should be processed at least once
-	if atomic.LoadInt32(&processCount) < 1 {
-		t.Errorf("Expected at least 1 processing, got %d", processCount)
-	}
-
-	// Verify store was updated
-	store.mu.Lock()
-	wasStored := store.processed[400]
-	store.mu.Unlock()
-
-	// The issue should have been marked in the store (via markProcessed in poller)
-	// Note: The poller marks issues as processed internally, and the store persists this
-	if !wasStored {
-		// This is expected behavior - the store is only updated when poller marks it
-		// The test verifies the callback is invoked and processing happens
-		t.Log("Note: Store not updated directly - this is expected as poller marks internally")
-	}
-}
-
-// mockProcessedStore implements ProcessedStore for testing
-type mockProcessedStore struct {
-	mu        sync.Mutex
-	processed map[int]bool
-}
-
-func (m *mockProcessedStore) Mark(source, repo, issueID string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if id, err := strconv.Atoi(issueID); err == nil {
-		m.processed[id] = true
-	}
-	return nil
-}
-
-func (m *mockProcessedStore) Unmark(source, repo, issueID string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if id, err := strconv.Atoi(issueID); err == nil {
-		delete(m.processed, id)
-	}
-	return nil
-}
-
-func (m *mockProcessedStore) IsProcessed(source, repo, issueID string) (bool, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	if id, err := strconv.Atoi(issueID); err == nil {
-		return m.processed[id], nil
-	}
-	return false, nil
-}
-
-func (m *mockProcessedStore) Load(source, repo string) (map[string]time.Time, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	result := make(map[string]time.Time)
-	for k := range m.processed {
-		result[strconv.Itoa(k)] = time.Now()
-	}
-	return result, nil
-}
