--- stress/memory_test.go
@@ -1,492 +1,0 @@
-package stress
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"runtime"
-	"strconv"
-	"strings"
-	"sync/atomic"
-	"testing"
-	"time"
-
-	"github.com/qf-studio/pilot/internal/adapters/github"
-	"github.com/qf-studio/pilot/internal/testutil"
-)
-
-// waitForProcessed busy-waits until get() reaches want, failing fast with a clear
-// message when timeout elapses instead of hanging until the whole-package test
-// timeout (600s). The unbounded version of this loop made these stress tests red
-// CI on unrelated PRs whenever dispatch was slow under load. TASK-353.
-func waitForProcessed(t *testing.T, get func() int64, want int64, timeout time.Duration) {
-	t.Helper()
-	deadline := time.Now().Add(timeout)
-	for get() < want {
-		if time.Now().After(deadline) {
-			t.Fatalf("timed out after %v waiting for processed count: got %d, want %d", timeout, get(), want)
-		}
-		time.Sleep(50 * time.Millisecond)
-	}
-}
-
-// TestMemory_NoUnboundedGrowth verifies memory doesn't grow unbounded during processing.
-func TestMemory_NoUnboundedGrowth(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping memory test in short mode")
-	}
-
-	const (
-		numIssues     = 100
-		maxConcurrent = 10
-		// Allow 50MB growth max - generous for test overhead
-		maxGrowthBytes = 50 * 1024 * 1024
-	)
-
-	// Force GC and baseline
-	runtime.GC()
-	time.Sleep(100 * time.Millisecond)
-
-	var initialMem runtime.MemStats
-	runtime.ReadMemStats(&initialMem)
-
-	var processedCount int64
-
-	issues := make([]*github.Issue, numIssues)
-	for i := 0; i < numIssues; i++ {
-		issues[i] = &github.Issue{
-			Number:    i + 1,
-			Title:     fmt.Sprintf("Memory Test Issue %d", i+1),
-			Body:      fmt.Sprintf("Body with some content for issue %d to simulate realistic payload size", i+1),
-			State:     "open",
-			Labels:    []github.Label{{Name: "pilot"}},
-			CreatedAt: time.Now().Add(-time.Duration(numIssues-i) * time.Minute),
-		}
-	}
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		// C6/TASK-346: paginate-aware mock — full list on page 1, empty after, so
-		// ListIssues pagination terminates instead of looping to maxPages. TASK-353.
-		if p := r.URL.Query().Get("page"); p != "" && p != "1" {
-			_, _ = w.Write([]byte(`[]`))
-			return
-		}
-		_ = json.NewEncoder(w).Encode(issues)
-	}))
-	defer server.Close()
-
-	client := github.NewClientWithBaseURL(testutil.FakeGitHubToken, server.URL)
-
-	poller, err := github.NewPoller(
-		client,
-		"owner/repo",
-		"pilot",
-		50*time.Millisecond,
-		github.WithMaxConcurrent(maxConcurrent),
-		github.WithOnIssue(func(ctx context.Context, issue *github.Issue) error {
-			// Simulate some work that allocates memory
-			_ = make([]byte, 1024) // 1KB allocation per issue
-			time.Sleep(10 * time.Millisecond)
-			atomic.AddInt64(&processedCount, 1)
-			return nil
-		}),
-	)
-	if err != nil {
-		t.Fatalf("Failed to create poller: %v", err)
-	}
-
-	// Track memory during processing
-	metrics := NewMetrics()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	go poller.Start(ctx)
-
-	// Sample memory periodically
-	done := make(chan struct{})
-	go func() {
-		ticker := time.NewTicker(100 * time.Millisecond)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-done:
-				return
-			case <-ticker.C:
-				metrics.SampleMemoryAndGoroutines()
-			}
-		}
-	}()
-
-	// Wait for processing
-	waitForProcessed(t, func() int64 { return atomic.LoadInt64(&processedCount) }, int64(numIssues), 25*time.Second)
-
-	close(done)
-	cancel()
-	poller.WaitForActive()
-
-	// Force GC to get accurate final memory
-	runtime.GC()
-	time.Sleep(100 * time.Millisecond)
-
-	var finalMem runtime.MemStats
-	runtime.ReadMemStats(&finalMem)
-
-	memGrowth := int64(finalMem.Alloc) - int64(initialMem.Alloc)
-
-	t.Logf("Memory stats:")
-	t.Logf("  Initial: %d MB", initialMem.Alloc/1024/1024)
-	t.Logf("  Peak: %d MB", metrics.GetPeakMemory()/1024/1024)
-	t.Logf("  Final: %d MB", finalMem.Alloc/1024/1024)
-	t.Logf("  Growth: %d bytes", memGrowth)
-	t.Logf("  Issues processed: %d", atomic.LoadInt64(&processedCount))
-
-	// Memory growth should be bounded
-	if memGrowth > maxGrowthBytes {
-		t.Errorf("Memory grew by %d bytes (%.2f MB), exceeds limit of %d bytes",
-			memGrowth, float64(memGrowth)/1024/1024, maxGrowthBytes)
-	}
-}
-
-// TestMemory_ProcessedMapGrowth verifies the processed map doesn't leak.
-func TestMemory_ProcessedMapGrowth(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping memory test in short mode")
-	}
-
-	const numIssues = 1000
-
-	issues := make([]*github.Issue, numIssues)
-	for i := 0; i < numIssues; i++ {
-		issues[i] = &github.Issue{
-			Number:    i + 1,
-			Title:     fmt.Sprintf("Issue %d", i+1),
-			State:     "open",
-			Labels:    []github.Label{{Name: "pilot"}},
-			CreatedAt: time.Now(),
-		}
-	}
-
-	// After the initial dispatch poll, return issues with pilot-in-progress label
-	// so the retry logic doesn't clear them from the processed map on subsequent ticks.
-	// recoverOrphanedIssues also makes a ListIssues call, so we track calls and
-	// return the dispatchable set only on the second call (first actual poll).
-	var pollCount int64
-	inProgressIssues := make([]*github.Issue, numIssues)
-	for i := 0; i < numIssues; i++ {
-		inProgressIssues[i] = &github.Issue{
-			Number:    i + 1,
-			Title:     fmt.Sprintf("Issue %d", i+1),
-			State:     "open",
-			Labels:    []github.Label{{Name: "pilot"}, {Name: "pilot-in-progress"}},
-			CreatedAt: time.Now(),
-		}
-	}
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		// Single-issue GET (e.g. /repos/owner/repo/issues/42): return one fresh
-		// issue so the poller's label re-check (GH-2341) doesn't decode a
-		// 1000-element array per dispatch.
-		if strings.Contains(r.URL.Path, "/issues/") {
-			parts := strings.Split(r.URL.Path, "/")
-			if n, perr := strconv.Atoi(parts[len(parts)-1]); perr == nil && n >= 1 && n <= numIssues {
-				_ = json.NewEncoder(w).Encode(issues[n-1])
-				return
-			}
-		}
-		// TASK-321 PR-4: parallel checkForNewIssues now runs the fresh-candidate
-		// merged-work guard, which probes /search/issues + /pulls per candidate.
-		// Answer "no merged work" cheaply so the guard is a fast no-op — otherwise
-		// it would decode the 1000-element list per issue and starve the 30s poll.
-		if strings.HasPrefix(r.URL.Path, "/search/issues") {
-			_, _ = w.Write([]byte(`{"total_count":0,"items":[]}`))
-			return
-		}
-		if strings.HasSuffix(r.URL.Path, "/pulls") {
-			_, _ = w.Write([]byte(`[]`))
-			return
-		}
-		// C6/TASK-346: ListIssues now paginates (per_page=100&page=N). Return the
-		// full list only on page 1 and an empty page after, so pagination terminates;
-		// otherwise the mock would return 1000 per page up to maxPages (50k items),
-		// starving the poll. TASK-353.
-		if p := r.URL.Query().Get("page"); p != "" && p != "1" {
-			_, _ = w.Write([]byte(`[]`))
-			return
-		}
-		n := atomic.AddInt64(&pollCount, 1)
-		if n <= 2 {
-			// Call 1: recoverOrphanedIssues, Call 2: first checkForNewIssues
-			_ = json.NewEncoder(w).Encode(issues)
-		} else {
-			_ = json.NewEncoder(w).Encode(inProgressIssues)
-		}
-	}))
-	defer server.Close()
-
-	client := github.NewClientWithBaseURL(testutil.FakeGitHubToken, server.URL)
-
-	var processedCount int64
-
-	poller, err := github.NewPoller(
-		client,
-		"owner/repo",
-		"pilot",
-		10*time.Millisecond,
-		github.WithExecutionMode(github.ExecutionModeParallel),
-		github.WithMaxConcurrent(20),
-		github.WithOnIssue(func(ctx context.Context, issue *github.Issue) error {
-			atomic.AddInt64(&processedCount, 1)
-			return nil
-		}),
-	)
-	if err != nil {
-		t.Fatalf("Failed to create poller: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	go poller.Start(ctx)
-
-	// Wait for all to be processed
-	waitForProcessed(t, func() int64 { return atomic.LoadInt64(&processedCount) }, int64(numIssues), 25*time.Second)
-
-	cancel()
-	poller.WaitForActive()
-
-	// Verify processed count matches
-	if count := poller.ProcessedCount(); count != numIssues {
-		t.Errorf("ProcessedCount() = %d, want %d", count, numIssues)
-	}
-
-	// Reset should clear processed map
-	poller.Reset()
-	if count := poller.ProcessedCount(); count != 0 {
-		t.Errorf("ProcessedCount() after Reset() = %d, want 0", count)
-	}
-}
-
-// TestMemory_LargePayloads tests handling of issues with large bodies.
-func TestMemory_LargePayloads(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping memory test in short mode")
-	}
-
-	const (
-		numIssues     = 20
-		bodySize      = 100 * 1024 // 100KB per issue body
-		maxConcurrent = 5
-	)
-
-	// Generate large issue bodies
-	largeBody := make([]byte, bodySize)
-	for i := range largeBody {
-		largeBody[i] = byte('a' + i%26)
-	}
-	bodyStr := string(largeBody)
-
-	issues := make([]*github.Issue, numIssues)
-	for i := 0; i < numIssues; i++ {
-		issues[i] = &github.Issue{
-			Number:    i + 1,
-			Title:     fmt.Sprintf("Large Issue %d", i+1),
-			Body:      bodyStr,
-			State:     "open",
-			Labels:    []github.Label{{Name: "pilot"}},
-			CreatedAt: time.Now(),
-		}
-	}
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		// mem-048: the poller's dispatch cycle also calls GetIssue (GH-2341 fresh
-		// label re-check) and the merged-work guard (GH-1983/GH-3269: search +
-		// branch lookup) for every candidate. Those routes must be served with
-		// correctly-shaped, cheap responses — otherwise they fall through to the
-		// full 2MB issue-list branch below, which fails to decode into the
-		// expected shape and multiplies large-payload traffic per issue on every
-		// poll tick (known pitfall: unserved routes in the poller poll/dispatch
-		// cycle stall the stress suite).
-		if strings.Contains(r.URL.Path, "/issues/") {
-			parts := strings.Split(r.URL.Path, "/")
-			if n, perr := strconv.Atoi(parts[len(parts)-1]); perr == nil && n >= 1 && n <= numIssues {
-				_ = json.NewEncoder(w).Encode(issues[n-1])
-				return
-			}
-		}
-		if strings.HasPrefix(r.URL.Path, "/search/issues") {
-			_, _ = w.Write([]byte(`{"total_count":0,"items":[]}`))
-			return
-		}
-		if strings.HasSuffix(r.URL.Path, "/pulls") {
-			_, _ = w.Write([]byte(`[]`))
-			return
-		}
-		// C6/TASK-346: paginate-aware mock — full list on page 1, empty after, so
-		// ListIssues pagination terminates instead of looping to maxPages. TASK-353.
-		if p := r.URL.Query().Get("page"); p != "" && p != "1" {
-			_, _ = w.Write([]byte(`[]`))
-			return
-		}
-		_ = json.NewEncoder(w).Encode(issues)
-	}))
-	defer server.Close()
-
-	client := github.NewClientWithBaseURL(testutil.FakeGitHubToken, server.URL)
-
-	var processedCount int64
-
-	poller, err := github.NewPoller(
-		client,
-		"owner/repo",
-		"pilot",
-		100*time.Millisecond,
-		github.WithMaxConcurrent(maxConcurrent),
-		github.WithOnIssue(func(ctx context.Context, issue *github.Issue) error {
-			// Verify body is intact
-			if len(issue.Body) != bodySize {
-				return fmt.Errorf("body size mismatch: got %d, want %d", len(issue.Body), bodySize)
-			}
-			time.Sleep(20 * time.Millisecond)
-			atomic.AddInt64(&processedCount, 1)
-			return nil
-		}),
-	)
-	if err != nil {
-		t.Fatalf("Failed to create poller: %v", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	go poller.Start(ctx)
-
-	// GH-3959: bounded wait (mirrors waitForProcessed used elsewhere in this
-	// file, TASK-353) — an unbounded busy-wait here let any transient dispatch
-	// slowness escalate into a full 10m package-timeout kill instead of a fast,
-	// clear failure.
-	waitForProcessed(t, func() int64 { return atomic.LoadInt64(&processedCount) }, int64(numIssues), 25*time.Second)
-
-	cancel()
-	poller.WaitForActive()
-
-	processed := atomic.LoadInt64(&processedCount)
-	if processed != int64(numIssues) {
-		t.Errorf("Processed %d issues, want %d", processed, numIssues)
-	}
-
-	t.Logf("Successfully processed %d issues with %dKB payloads each", numIssues, bodySize/1024)
-}
-
-// TestMemory_RepeatedStartStop tests for memory leaks across multiple start/stop cycles.
-func TestMemory_RepeatedStartStop(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping memory test in short mode")
-	}
-
-	const (
-		numCycles      = 10
-		issuesPerCycle = 10
-	)
-
-	runtime.GC()
-	time.Sleep(100 * time.Millisecond)
-
-	var initialMem runtime.MemStats
-	runtime.ReadMemStats(&initialMem)
-
-	baseIssues := make([]*github.Issue, issuesPerCycle)
-	for i := 0; i < issuesPerCycle; i++ {
-		baseIssues[i] = &github.Issue{
-			Number:    i + 1,
-			Title:     fmt.Sprintf("Issue %d", i+1),
-			State:     "open",
-			Labels:    []github.Label{{Name: "pilot"}},
-			CreatedAt: time.Now(),
-		}
-	}
-
-	for cycle := 0; cycle < numCycles; cycle++ {
-		// Create fresh issues for each cycle
-		issues := make([]*github.Issue, issuesPerCycle)
-		for i := 0; i < issuesPerCycle; i++ {
-			issues[i] = &github.Issue{
-				Number:    (cycle * issuesPerCycle) + i + 1,
-				Title:     fmt.Sprintf("Cycle %d Issue %d", cycle+1, i+1),
-				State:     "open",
-				Labels:    []github.Label{{Name: "pilot"}},
-				CreatedAt: time.Now(),
-			}
-		}
-
-		server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Header().Set("Content-Type", "application/json")
-			// C6/TASK-346: paginate-aware mock — full list on page 1, empty after.
-			if p := r.URL.Query().Get("page"); p != "" && p != "1" {
-				_, _ = w.Write([]byte(`[]`))
-				return
-			}
-			_ = json.NewEncoder(w).Encode(issues)
-		}))
-
-		client := github.NewClientWithBaseURL(testutil.FakeGitHubToken, server.URL)
-
-		var processedCount int64
-
-		poller, err := github.NewPoller(
-			client,
-			"owner/repo",
-			"pilot",
-			10*time.Millisecond,
-			github.WithMaxConcurrent(5),
-			github.WithOnIssue(func(ctx context.Context, issue *github.Issue) error {
-				atomic.AddInt64(&processedCount, 1)
-				return nil
-			}),
-		)
-		if err != nil {
-			t.Fatalf("Cycle %d: failed to create poller: %v", cycle, err)
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		go poller.Start(ctx)
-
-		// Wait for processing
-		deadline := time.After(5 * time.Second)
-		for atomic.LoadInt64(&processedCount) < int64(issuesPerCycle) {
-			select {
-			case <-deadline:
-				t.Fatalf("Cycle %d timed out", cycle)
-			default:
-				time.Sleep(10 * time.Millisecond)
-			}
-		}
-
-		cancel()
-		poller.WaitForActive()
-		server.Close()
-
-		// Force GC between cycles
-		runtime.GC()
-	}
-
-	// Final cleanup
-	time.Sleep(500 * time.Millisecond)
-	runtime.GC()
-	time.Sleep(100 * time.Millisecond)
-
-	var finalMem runtime.MemStats
-	runtime.ReadMemStats(&finalMem)
-
-	// Allow 10MB growth for test overhead
-	maxGrowth := uint64(10 * 1024 * 1024)
-	if finalMem.Alloc > initialMem.Alloc+maxGrowth {
-		t.Errorf("Memory leak detected: initial=%dMB, final=%dMB, growth=%dMB",
-			initialMem.Alloc/1024/1024, finalMem.Alloc/1024/1024,
-			(finalMem.Alloc-initialMem.Alloc)/1024/1024)
-	}
-
-	t.Logf("Completed %d cycles, memory: initial=%dMB, final=%dMB",
-		numCycles, initialMem.Alloc/1024/1024, finalMem.Alloc/1024/1024)
-}
